srcs/requirements/django/app/requirements.txt
# requirements.txt

# Framework
## Django
django==5.1

# Forms
## Crispy and Boostrap pack
django-crispy-forms==2.3
crispy-bootstrap5==2024.2

# Websockets
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# ASGI tools
asgiref==3.8.1

# Database
## PostgreSQL 
### psycopg is the PostgreSQL adapter for Python
psycopg2-binary==2.9.9
sqlparse==0.5.1

# TLS and HTTP/2 support
twisted[tls,http2]==24.7.0
## Fast HTTP Request library
requests==2.32.3
# Python Imaging Library
pillow==10.4.0

# Third Party Authentication
## Social account authentication (Google)
django-allauth==64.0.0
## JSON Web Tokens
PyJWT==2.9.0